pyproject.toml
[project]
name = "shisad"
version = "0.6.0"
description = "Security-first AI agent daemon"
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "agent-client-protocol==0.8.1",
    "click>=8.1,<9",
    "pyyaml>=6.0,<7",
    "cryptography>=46.0.6,<47",
    "loguru>=0.7,<1",
]

[project.urls]
Homepage = "https://github.com/shisa-ai/shisad"
Repository = "https://github.com/shisa-ai/shisad"
Issues = "https://github.com/shisa-ai/shisad/issues"
Changelog = "https://github.com/shisa-ai/shisad/blob/main/CHANGELOG.md"

[project.scripts]
shisad = "shisad.cli.main:cli"

[build-system]
requires = ["hatchling==1.29.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shisad"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "requires_cap_net_admin: tests that require CAP_NET_ADMIN capability",
    "live_smoke: minimal high-signal live-model behavioral coverage",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[dependency-groups]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "types-pyyaml>=6.0.12.20250915",
    "textual>=0.89,<1",
]
security-runtime = [
    "yara-python>=4.5,<5",
]
coverage = [
    "pytest-cov>=5.0,<7",
]
channels-runtime = [
    "matrix-nio[e2e]>=0.25,<0.26",
    "discord.py>=2.4,<3",
    "python-telegram-bot>=21.6,<22",
    "slack-bolt>=1.21,<2",
    "slack-sdk>=3.33,<4",
]
